Move StorageContext to storage/context
