+++ Cargo.toml
@@ -46,6 +46,7 @@ members = [
     "src/query/legacy-expression",
     "src/query/legacy-planners",
     "src/query/settings",
+    "src/query/storages/context",
     "src/query/storages/fuse",
     "src/query/storages/fuse-meta",
     "src/query/storages/hive",
