+++ src/query/service/src/storages/mod.rs
@@ -25,13 +25,12 @@ pub mod system;
 
 pub use common_catalog::table::NavigationPoint;
 pub use common_catalog::table::TableStatistics;
+pub use common_storages_context::StorageContext;
 pub use common_storages_preludes::information_schema;
 pub use common_storages_preludes::memory;
 pub use common_storages_preludes::null;
 pub use common_storages_preludes::random;
 pub use common_storages_preludes::view;
-use common_storages_util::storage_context;
-pub use storage_context::StorageContext;
 pub use storage_factory::StorageCreator;
 pub use storage_factory::StorageDescription;
 pub use storage_factory::StorageFactory;
