+++ src/query/storages/context/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "common-storages-context"
+version = { workspace = true }
+authors = { workspace = true }
+license = { workspace = true }
+publish = { workspace = true }
+edition = { workspace = true }
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+doctest = false
+test = false
+
+[dependencies]
+common-datablocks = { path = "../../datablocks" }
+
+parking_lot = "0.12.1"
