+++ src/query/storages/context/src/lib.rs
@@ -1,3 +1,5 @@
+// Copyright 2022 Datafuse Labs.
+//
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -12,7 +14,7 @@
 
 #![deny(unused_crate_dependencies)]
 
-pub mod cached_reader;
-pub mod file_meta_data_reader;
-pub mod table_option_keys;
-pub mod table_storage_prefix;
+#[allow(clippy::module_inception)]
+mod context;
+
+pub use context::StorageContext;
