+++ src/query/storages/fuse/src/fuse_table.rs
@@ -45,7 +45,7 @@ use common_sharing::create_share_table_operator;
 use common_storage::init_operator;
 use common_storage::DataOperator;
 use common_storage::ShareTableConfig;
-use common_storages_util::storage_context::StorageContext;
+use common_storages_context::StorageContext;
 use common_storages_util::table_storage_prefix::table_storage_prefix;
 use opendal::Operator;
 use uuid::Uuid;
