+++ src/query/storages/preludes/Cargo.toml
@@ -27,6 +27,7 @@ common-pipeline-core = { path = "../../pipeline/core" }
 common-pipeline-sinks = { path = "../../pipeline/sinks" }
 common-pipeline-sources = { path = "../../pipeline/sources" }
 common-pipeline-transforms = { path = "../../pipeline/transforms" }
+common-storages-context = { path = "../context" }
 common-storages-util = { path = "../util" }
 common-users = { path = "../../users" }
