+++ src/query/storages/util/Cargo.toml
@@ -15,11 +15,9 @@ test = false
 common-arrow = { path = "../../../common/arrow" }
 common-cache = { path = "../../../common/cache" }
 common-catalog = { path = "../../../common/../query/catalog" }
-common-datablocks = { path = "../../datablocks" }
 common-exception = { path = "../../../common/exception" }
 common-fuse-meta = { path = "../fuse-meta" }
 
 async-trait = { version = "0.1.57", package = "async-trait-fn" }
 once_cell = "1.15.0"
 opendal = { version = "0.19", features = ["layers-retry"] }
-parking_lot = "0.12.1"
